Fix to support remote ops
Remote ops require Python packages --- any modules not in a package
cannot be included in the generated wheel. This is a constraint
imposed by the Python packaging system, which relies on top-level
Python packages (dirs with __init__.py files) to implement namespaces.

To verify the fix, run TESTS.md as a Guild test:

    $ guild check -nt TESTS.md

The tests require that sshd be running on the current system such that
`ssh localhost true` exits without error. This is verified by the
tests.